requirements.txt
tensorflow==2.10.0
numpy==1.21.0
matplotlib==3.4.2
pandas==1.3.0
opencv-python==4.5.3.56
scikit-learn==0.24.2
Pillow==8.3.1
streamlit==0.84.0
kagglehub==0.1.0